--- a/pkg/kubelet/apis/pluginregistration/v1beta1/BUILD
+++ b/pkg/kubelet/apis/pluginregistration/v1beta1/BUILD
@@ -15,8 +15,9 @@ go_library(
     deps = [
         "//vendor/github.com/gogo/protobuf/gogoproto:go_default_library",
         "//vendor/github.com/gogo/protobuf/proto:go_default_library",
-        "//vendor/golang.org/x/net/context:go_default_library",
         "//vendor/google.golang.org/grpc:go_default_library",
+        "//vendor/google.golang.org/grpc/codes:go_default_library",
+        "//vendor/google.golang.org/grpc/status:go_default_library",
     ],
 )