--- a/staging/src/k8s.io/apimachinery/pkg/apis/meta/v1/time_proto.go
+++ b/staging/src/k8s.io/apimachinery/pkg/apis/meta/v1/time_proto.go
@@ -75,7 +75,7 @@ func (m *Time) Unmarshal(data []byte) error {
 	return nil
 }
 
-// Marshal implements the protobuf marshalling interface.
+// Marshal implements the protobuf marshaling interface.
 func (m *Time) Marshal() (data []byte, err error) {
 	if m == nil || m.Time.IsZero() {
 		return nil, nil
@@ -83,10 +83,18 @@ func (m *Time) Marshal() (data []byte, err error) {
 	return m.ProtoTime().Marshal()
 }
 
-// MarshalTo implements the protobuf marshalling interface.
+// MarshalTo implements the protobuf marshaling interface.
 func (m *Time) MarshalTo(data []byte) (int, error) {
 	if m == nil || m.Time.IsZero() {
 		return 0, nil
 	}
 	return m.ProtoTime().MarshalTo(data)
+}
+
+// MarshalToSizedBuffer implements the protobuf reverse marshaling interface.
+func (m *Time) MarshalToSizedBuffer(data []byte) (int, error) {
+	if m == nil || m.Time.IsZero() {
+		return 0, nil
+	}
+	return m.ProtoTime().MarshalToSizedBuffer(data)
 }