--- a/staging/src/k8s.io/cri-api/pkg/apis/runtime/v1alpha2/BUILD
+++ b/staging/src/k8s.io/cri-api/pkg/apis/runtime/v1alpha2/BUILD
@@ -13,8 +13,9 @@ go_library(
         "//vendor/github.com/gogo/protobuf/gogoproto:go_default_library",
         "//vendor/github.com/gogo/protobuf/proto:go_default_library",
         "//vendor/github.com/gogo/protobuf/sortkeys:go_default_library",
-        "//vendor/golang.org/x/net/context:go_default_library",
         "//vendor/google.golang.org/grpc:go_default_library",
+        "//vendor/google.golang.org/grpc/codes:go_default_library",
+        "//vendor/google.golang.org/grpc/status:go_default_library",
     ],
 )