--- a/staging/src/k8s.io/metrics/go.mod
+++ b/staging/src/k8s.io/metrics/go.mod
@@ -5,7 +5,7 @@ module k8s.io/metrics
 go 1.12
 
 require (
-	github.com/gogo/protobuf v1.0.0
+	github.com/gogo/protobuf v0.0.0-20190410021324-65acae22fc9
 	github.com/stretchr/testify v1.3.0
 	k8s.io/api v0.0.0
 	k8s.io/apimachinery v0.0.0