… go.sum
… pkg/kubelet/apis/deviceplugin/v1alpha/api.pb.go
… pkg/kubelet/apis/deviceplugin/v1beta1/api.pb.go
… pkg/kubelet/apis/pluginregistration/v1alpha1/api.pb.go
… pkg/kubelet/apis/pluginregistration/v1beta1/api.pb.go
… pkg/kubelet/pluginmanager/pluginwatcher/example_plugin_apis/v1beta1/api.pb.go
… pkg/kubelet/pluginmanager/pluginwatcher/example_plugin_apis/v1beta2/api.pb.go
… staging/src/k8s.io/api/admission/v1beta1/generated.pb.go
… staging/src/k8s.io/api/admissionregistration/v1/generated.pb.go
… staging/src/k8s.io/api/admissionregistration/v1beta1/generated.pb.go
… staging/src/k8s.io/api/apps/v1/generated.pb.go
… staging/src/k8s.io/api/apps/v1beta1/generated.pb.go
… staging/src/k8s.io/api/apps/v1beta2/generated.pb.go
… staging/src/k8s.io/api/auditregistration/v1alpha1/generated.pb.go
… staging/src/k8s.io/api/authentication/v1/generated.pb.go
… staging/src/k8s.io/api/authentication/v1beta1/generated.pb.go
… staging/src/k8s.io/api/authorization/v1/generated.pb.go
… staging/src/k8s.io/api/authorization/v1beta1/generated.pb.go
… staging/src/k8s.io/api/autoscaling/v1/generated.pb.go
… staging/src/k8s.io/api/autoscaling/v2beta1/generated.pb.go
… staging/src/k8s.io/api/autoscaling/v2beta2/generated.pb.go
… staging/src/k8s.io/api/batch/v1/generated.pb.go
… staging/src/k8s.io/api/batch/v1beta1/generated.pb.go
… staging/src/k8s.io/api/batch/v2alpha1/generated.pb.go
… staging/src/k8s.io/api/certificates/v1beta1/generated.pb.go
… staging/src/k8s.io/api/coordination/v1/generated.pb.go
… staging/src/k8s.io/api/coordination/v1beta1/generated.pb.go
… staging/src/k8s.io/api/core/v1/generated.pb.go
… staging/src/k8s.io/api/events/v1beta1/generated.pb.go
… staging/src/k8s.io/api/extensions/v1beta1/generated.pb.go