This is trying a new version of gogo/protobuf serialization that goes backward. It's still a WIP, especially since the code is not merged in that repo, so we're replacing in go.mod to point to the pull-request instead. Let's see if this works.

**What type of PR is this?**
/kind bug

**What this PR does / why we need it**:

**Which issue(s) this PR fixes**:
Fixes #76219

**Special notes for your reviewer**:

**Does this PR introduce a user-facing change?**:
```release-note
Update gogo/protobuf to serialize backward, as to get better performance on deep objects.
```