@mattjmcnaughton 2019-05-03
Very cool stuff @apelisse :) I'm curious about your test plans for this - how do you plan to verify the performance improvement? In this pr or in a separate pr?

Thanks again for all your work on this :)