@apelisse 2019-05-03
> how do you plan to verify the performance improvement? In this pr or in a separate pr?

We are planning to test that PR on the scalability cluster with 5000 nodes and see what happens!