@apelisse 2019-06-07
@smarterclayton @liggitt This is ready for review, not ready to be merged. We still haven' t merged the code up-stream, but I'd like to know that this is something we would want to move forward with.

Upstream PR: https://github.com/gogo/protobuf/pull/560