@wojtek-t 2019-06-24
> @krzysied @wojtek-t This is the PR that pulls in the marshaling fix.

I was OOO for the last 3 weeks - sorry for delay.
So what are the benchmark results with the current changes?