@apelisse 2019-06-27
OK I've had the green light to rebase the code and re-run the tests. I'll do that and we'll see what happens.

Review-wise, I think I need opinions, especially on the kubernetes changes in that PR that it implies.