@apelisse 2019-06-27
> Review-wise, I think I need opinions, especially on the kubernetes changes in that PR that it implies.

It means this commit in particular: https://github.com/kubernetes/kubernetes/pull/77355/commits/0e746bde3e9fcca802da98a34d3d4f91b8feff4d