@smarterclayton 2019-07-08
The custom ones look ok (although I trust the round trip tests more than my code review of hand written protobuf encoding) :)