@smarterclayton 2019-07-10
Can you add a new benchmark for proto encode and decode that create PodLists of size 10, 100, 1000 and show before and after?