@apelisse 2019-07-10
```
benchmark                                           old ns/op  new ns/op   delta
BenchmarkEncodeProtobufGeneratedMarshalList10       64109      43583       -32.02%
BenchmarkEncodeProtobufGeneratedMarshalList100      881276     546877      -37.94%
BenchmarkEncodeProtobufGeneratedMarshalList1000     6960907    4333797     -37.74%

benchmark                                           old allocs  new allocs   delta
BenchmarkEncodeProtobufGeneratedMarshalList10       71          51           -28.17%
BenchmarkEncodeProtobufGeneratedMarshalList100      546         442          -19.05%
BenchmarkEncodeProtobufGeneratedMarshalList1000     5210        4190         -19.58%

benchmark                                           old bytes  new bytes   delta
BenchmarkEncodeProtobufGeneratedMarshalList10       11728      11376       -3.00%
BenchmarkEncodeProtobufGeneratedMarshalList100      119794     117967      -1.53%
BenchmarkEncodeProtobufGeneratedMarshalList1000     1183885    1165941     -1.52%
```