@apelisse 2019-07-24
This is not a WIP anymore. The code has been merged upstream and can now be used in Kubernetes.
It still has about 40% speed improvement on podlist serializations, and 30% on nodelists. 

/priority important-soon