@wojtek-t 2019-07-25
@apelisse - verify godeps is failing and that seems like a legitimate failure (though seems trivial to fix)