@apelisse 2019-07-25
> verify godeps is failing and that seems like a legitimate failure (though seems trivial to fix)

Yes, I'll fix this soon :-)