@BenTheElder 2019-07-25
/assign @yujuhong @Random-Liu 
PTAL for cri-api (the change is just proto optimization, not API changes)