@apelisse 2019-07-30
Hard to know what really is going on yet, but that's certainly related: https://github.com/gogo/protobuf/issues/589